+++ Controller/Payment/Callback/Error.php
@@ -11,27 +11,20 @@ use Magento\Framework\App\RequestInterface;
  */
 class Error extends \Magento\Framework\App\Action\Action implements \Magento\Framework\App\CsrfAwareActionInterface
 {
-    /**
-     * @var ModuleConfig
-     */
-//    private $moduleConfig;
-    
     private $readerWriter;
 
     /**
      * Error constructor.
      *
-     * @param Context          $context
-     * @param ModuleConfig     $moduleConfig
+     * @param Context       $context
+     * @param ReaderWriter  $readerWriter
      */
     public function __construct(
         \Magento\Framework\App\Action\Context $context,
-        //        \Nuvei\Checkout\Model\Config $moduleConfig,
         \Nuvei\Checkout\Model\ReaderWriter $readerWriter
     ) {
         parent::__construct($context);
 
-//        $this->moduleConfig = $moduleConfig;
         $this->readerWriter = $readerWriter;
     }